+++ src/api/projects/merge_requests/notes/awards.rs
@@ -8,8 +8,14 @@
 //!
 //! These endpoints are used for querying project merge request note awards.
 
+mod award;
 mod awards;
 mod create;
+mod delete;
+
+pub use self::award::MergeRequestNoteAward;
+pub use self::award::MergeRequestNoteAwardBuilder;
+pub use self::award::MergeRequestNoteAwardBuilderError;
 
 pub use self::awards::MergeRequestNoteAwards;
 pub use self::awards::MergeRequestNoteAwardsBuilder;
@@ -18,3 +24,7 @@ pub use self::awards::MergeRequestNoteAwardsBuilderError;
 pub use self::create::CreateMergeRequestNoteAward;
 pub use self::create::CreateMergeRequestNoteAwardBuilder;
 pub use self::create::CreateMergeRequestNoteAwardBuilderError;
+
+pub use self::delete::DeleteMergeRequestNoteAward;
+pub use self::delete::DeleteMergeRequestNoteAwardBuilder;
+pub use self::delete::DeleteMergeRequestNoteAwardBuilderError;